backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
groq==0.13.0
sentence-transformers==3.3.1
numpy==2.2.1
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.20
scikit-learn==1.6.1
sqlite-utils==3.38
